WIP: bind dnsmasq to lan0 device
does this help us with dnsmasq not starting until lan0 is up
does it also help us if lan0 goes down, that this service goes away?

is this really an improvement?  cause in the past if lan0 goes away
i have had to reboot the box to get plugin data out of the node
i think k3s needs to be restarted too...