SAGE-1543: configure the NX agent for 10W / 4 core operating mode
It was identified that the NX agent was being "power starved" when
stressed at 15W / 6 core mode when being powered over PoE from the
network swtich (the default and very common mode). Therefore, we are
setting the default behavior for the NX agent to be 10W / 4 core mode
to ensure it doesn't reboot due to lack of wattage during operation.