SAGE-1619: resolve missing 'cuda' mounts in Docker/K3S pods
The new version of libnvidia-container (and related tools) causes
for *only* the the "base files" (l4t.csv, drivers.csv, devices.csv)
to be processed by the NVidia runtime instead of *.csv.
Therefore, create our own "waggle" runtime mount file and symlink
"devices.csv" to it, to re-enable all the mounts to be processed like
it was done on earlier versions of libnvidia-container.

Seed the waggle specific csv file with the contents of cuda.csv (source:
nvidia-container-csv-cuda=10.2.89-1).